+++ test/clj/rems/application/test_commands.clj
@@ -1594,7 +1594,7 @@
                             :comment ""}
                            injections))))))
 
-(deftest test-promote-to-applicant
+(deftest test-change-applicant
   (let [application (apply-events nil
                                   [dummy-created-event
                                    {:event/type :application.event/submitted
@@ -1614,35 +1614,35 @@
               :application/applicant {:userid "member1"}
               :application/comment ""}
              (ok-command application
-                         {:type :application.command/promote-to-applicant
+                         {:type :application.command/change-applicant
                           :actor handler-user-id
                           :member {:userid "member1"}
                           :comment ""}))))
     (testing "applicant can't promote"
       (is (= {:errors [{:type :forbidden}]}
              (fail-command application
-                           {:type :application.command/promote-to-applicant
+                           {:type :application.command/change-applicant
                             :actor applicant-user-id
                             :member {:userid "member1"}
                             :comment ""}))))
     (testing "member can't promote themself"
       (is (= {:errors [{:type :forbidden}]}
              (fail-command application
-                           {:type :application.command/promote-to-applicant
+                           {:type :application.command/change-applicant
                             :actor "member1"
                             :member {:userid "member1"}
                             :comment ""}))))
     (testing "handler can't promote non-member"
       (is (= {:errors [{:type :user-not-member :user {:userid "unknown"}}]}
              (fail-command application
-                           {:type :application.command/promote-to-applicant
+                           {:type :application.command/change-applicant
                             :actor handler-user-id
                             :member {:userid "unknown"}
                             :comment ""}))))
     (testing "handler can't promote current applicant"
       (is (= {:errors [{:type :user-not-member :user {:userid "applicant"}}]}
              (fail-command application
-                           {:type :application.command/promote-to-applicant
+                           {:type :application.command/change-applicant
                             :actor handler-user-id
                             :member {:userid applicant-user-id}
                             :comment ""}))))))